requirements.txt
appdirs==1.4.3
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
cssselect==1.0.3
Django==2.2
django-filter==2.0.0
django-otp==0.5.1
djangorestframework==3.9.0
fake-useragent==0.1.11
httplib2==0.12.0
idna==2.8
lxml==4.2.5
Markdown==3.0.1
parse==1.9.0
Pillow==5.3.0
pycryptodome==3.7.2
pyee==5.0.0
PyMySQL==0.9.3
pyppeteer==0.0.25
pyquery==1.4.0
pytz==2018.7
requests==2.21.0
requests-html==0.9.0
six==1.12.0
tqdm==4.28.1
urllib3==1.24.1
w3lib==1.19.0
websockets==7.0
celery==4.3.0
redis==3.3.8
PyExecJS==1.5.1
